requirements.txt
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
deprecation==2.1.0
exceptiongroup==1.2.2
grpcio==1.66.1
idna==3.8
iniconfig==2.0.0
iso8601==2.1.0
numpy==2.1.0
packaging==24.1
pandas==2.2.2
pluggy==1.5.0
prettytable==3.11.0
protobuf==4.25.4
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
tinkoff==0.1.1
tinkoff-invest==1.0.5
tinkoff-investments==0.2.0b105
tomli==2.0.1
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
wcwidth==0.2.13
websocket-client==1.8.0
